Added Dark/Light UI version 1
Added the dark/light indicator to the HUD. No numbers on it yet though - I'm gonna wait until I implement that functionality into the mechanic itself before I do that. For now, it accurately displays which type you are in. To cycle through, you can press O to go left and P to go right.