Test Parallax Update
Using a copied component that's present but unused on each tileset in the Unity Hierarchy, I was able to create a small example of a parallax effect on the building background. I tried making a custom version of said parallax to work with the virtual camera instead, but it didn't look correct to me so it's unused. The script is still there if anyone wants to try using it.