Fixed player moving during cutscene bug and pushing MiniMapTest scene, and the process I took to making it.
The player's velocity is now zeroed when the cutscene happens to prevent them from moving. However, if they are walking the SFX still plays.

The MiniMapTest scene is where I was testing out the tutorials that I researched, figuring out the process for making the mini map and the extra tile palette that we'd need.