Added First Pass of Save/Load Feature
-- Added saving at statues. The following information is saved:
- position of the statue being saved at
- max energy
- max health
- status of unlocked abilities
- object status of the following objects
   - forced encounter managers
   - fuse boxes
   - laser receivers
   - switches

This information is loaded upon entering the StartLevel scene. It can also be loaded from the pause menu under the "Restart from Checkpoint" option. I still need to make it work across scenes properly, so that will be in the update alongside adding more UI. I will also add encryption in the next update to prevent the player from tampering with the data.